+++ DistFiles/ReleaseNotes.md
@@ -46,6 +46,10 @@ back-up plan.)
 
 # Release Notes
 
+## August 17 2020 Glyssen 1.5.0
+Improved reference text logic when matching to reporting clauses ("he said" blocks) to prevent duplication.
+English and Russian reference texts updated to match v. 101 of the FCBH New Testament Director's Guide.
+
 ## June 16 2020 Glyssen 1.3.13
 Switched localization to use XLIFF format instead of TMX. More (albeit minimal) UI languages available. Translations can be done via crowdin.com.